* Fulcum App Boiler copied into Fulcrum App * Fulcrum App Boiler deleted with all related files
