--- src/senaite/instruments/instruments/fulcrum/fulcrumboilerapp/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# This file is part of SENAITE.INSTRUMENTS
-#
-# Copyright 2019 by it's authors.
