--- src/senaite/instruments/instruments/fulcrum/fulcrumboilerapp/fulcrumboilerapp_import.pt
@@ -1,20 +1,0 @@
-<p></p>
-<label for='instrument_results_file'>File</label>&nbsp;
-<input type="file" name="instrument_results_file" id="instrument_results_file"/>&nbsp;&nbsp;
-<p></p>
-<h3>Advanced options</h3>
-<table cellpadding="0" cellspacing="0">
-    <tr>
-        <td><label for="results_override">Results override</label></td>
-        <td>
-            <select name="results_override" id="results_override">
-                <option value="nooverride">Don't override results</option>
-                <option value="override">Override non-empty results</option>
-                <option value="overrideempty">Override non-empty results (also with empty)</option>
-            </select>
-        </td>
-    </tr>
-</table>
-<p></p>
-<input name="firstsubmit" type="submit" value="Submit" i18n:attributes="value"/>
-<p></p>
